+++ hw_06_jpql/src/main/java/ru/otus/marchenko/repositories/JpaBookRepository.java
@@ -34,7 +34,7 @@ public class JpaBookRepository implements BookRepository {
     }
 
     @Override
-    public Optional<Book> findCopyByTitleAndAuthor(String title, Long authorId) {
+    public Optional<Book> findByTitleAndAuthor(String title, Long authorId) {
         EntityGraph<?> entityGraph = em.getEntityGraph("book-author-genres-entity-graph");
         TypedQuery<Book> query = em.createQuery(
                 "select b from Book b where b.title = :title and b.author.id = :authorId"
@@ -57,8 +57,14 @@ public class JpaBookRepository implements BookRepository {
 
     @Override
     public Book deleteById(long id) {
-        Book book = findById(id).orElseThrow(() -> new EntityNotFoundException("Book not found"));
-        em.remove(book);
+        EntityGraph<?> entityGraph = em.getEntityGraph("book-author-genres-entity-graph");
+        Map<String, Object> properties = new HashMap<>();
+        properties.put(FETCH.getKey(), entityGraph);
+        Book book = Optional.ofNullable(em.find(Book.class, id, properties))
+                .orElseThrow(() -> new EntityNotFoundException("Book not found"));
+        if(book != null) {
+            em.remove(book);
+        }
         return book;
     }
 }
